Use a mixin to implement extensibility.
Move `token_auto_auth` to util.
